Saving vorticity and grad(vorticity) to .dat files
I've replaced the original code with this, as we no longer need them as
auxiliary variables now. Tested on a toy problem, Paraview still seems
to correctly obtain the omega. However, a big plus is that we can now
also plot omega with Python! :) Furthermore, this allows us to perform
integrals using yt, so now we have a benchmark to compare methods with
each other.