Join simulation setups and add omega-AMR
I've removed the directory with compressible case simulation files, as
it is a lot harder to maintain the 2 simulation setups (things keep
changing, so it's useful to have only 1 file). On top of settings.par
there is now a small "guide" how to change back to incompressible.

Also, I'm not quite happy with omega-AMR yet, as it feels a bit
hacky right now (see mod_usr/specialrefine_grid).